Move graph_tracking.hpp to tt-metalium internal API
Move tt_metal/api/tt-metalium/graph_tracking.hpp to
tt_metal/api/tt-metalium/internal/ and update all include paths and
the sources.cmake header list accordingly.
