+++ tests/ttnn/unit_tests/gtests/test_graph_basic.cpp
@@ -18,7 +18,7 @@
 #include <tuple>
 #include <vector>
 
-#include <tt-metalium/graph_tracking.hpp>
+#include <tt-metalium/internal/graph_tracking.hpp>
 #include "gtest/gtest.h"
 #include <tt-metalium/shape.hpp>
 #include "ttnn/graph/graph_consts.hpp"
