+++ tests/ttnn/unit_tests/gtests/test_graph_capture_arguments_transpose.cpp
@@ -7,7 +7,7 @@
 #include <map>
 #include <vector>
 
-#include <tt-metalium/graph_tracking.hpp>
+#include <tt-metalium/internal/graph_tracking.hpp>
 #include "gtest/gtest.h"
 #include <tt-metalium/shape.hpp>
 #include "ttnn/graph/graph_processor.hpp"
