+++ tests/ttnn/unit_tests/gtests/test_graph_capture_arguments_untilize_with_unpadding.cpp
@@ -6,7 +6,7 @@
 #include <algorithm>
 #include <vector>
 
-#include <tt-metalium/graph_tracking.hpp>
+#include <tt-metalium/internal/graph_tracking.hpp>
 #include "gtest/gtest.h"
 #include <tt-metalium/shape.hpp>
 #include "ttnn/graph/graph_processor.hpp"
